tests run by buildr
