+++ buildfile
@@ -32,11 +32,11 @@ define 'assignments', :layout => ass_layout do
   #   doc.using :windowtitle => "Abandon All Hope", :private => true
   # end
 
-  define 'pset2' do
-    package :javadoc
-    javadoc sources
-    doc.using :windowtitle => "Abandon All Hope", :private => true
-  end
+  # define 'pset2' do
+  #   package :javadoc
+  #   javadoc sources
+  #   doc.using :windowtitle => "Abandon All Hope", :private => true
+  # end
 
   task :default => [:build, :doc]
